simplified normalization

+++ mnist-cnn.py
@@ -4,8 +4,8 @@ from keras.layers import Conv2D, MaxPooling2D, Flatten, Dense, Dropout
 # Load MNIST dataset -----------------------------------------------------------
 (X_train, y_train), (X_test, y_test) = keras.datasets.mnist.load_data()
 # Preprocessing ----------------------------------------------------------------
-X_train = X_train.reshape(-1, 28, 28, 1).astype('float32') / 255
-X_test = X_test.reshape(-1, 28, 28, 1).astype('float32') / 255
+X_train = X_train.reshape(-1, 28, 28, 1) / 255
+X_test = X_test.reshape(-1, 28, 28, 1) / 255
 # Define the model -------------------------------------------------------------
 model = keras.models.Sequential()
